+++ packages/remix-architect/server.ts
@@ -1,6 +1,6 @@
-// This has been added as a global in node 15+
-import AbortController from "abort-controller";
 import {
+  // This has been added as a global in node 15+
+  AbortController,
   Headers as NodeHeaders,
   Request as NodeRequest,
   formatServerError
