+++ packages/remix-node/formData.ts
@@ -1,4 +1,3 @@
-import Busboy from "busboy";
 import type { Readable } from "stream";
 
 export type UploadHandlerArgs = {
@@ -12,92 +11,6 @@ export type UploadHandlerArgs = {
 export type UploadHandler = (
   args: UploadHandlerArgs
 ) => Promise<string | File | undefined>;
-
-export async function parseFormData(
-  contentType: string,
-  stream: Readable,
-  abortController?: AbortController,
-  uploadHandler?: UploadHandler
-) {
-  let formData = new NodeFormData();
-  let fileWorkQueue: Promise<void>[] = [];
-
-  await new Promise<void>(async (resolve, reject) => {
-    let busboy = new Busboy({
-      highWaterMark: 2 * 1024 * 1024,
-      headers: {
-        "content-type": contentType
-      }
-    });
-
-    let aborted = false;
-    function abort(error?: Error) {
-      if (aborted) return;
-      aborted = true;
-
-      stream.unpipe();
-      stream.removeAllListeners();
-      busboy.removeAllListeners();
-
-      abortController?.abort();
-      reject(error || new Error("failed to parse form data"));
-    }
-
-    busboy.on("field", (name, value) => {
-      formData.append(name, value);
-    });
-
-    busboy.on("file", (name, filestream, filename, encoding, mimetype) => {
-      if (uploadHandler && filename) {
-        fileWorkQueue.push(
-          (async () => {
-            try {
-              let value = await uploadHandler({
-                name,
-                stream: filestream,
-                filename,
-                encoding,
-                mimetype
-              });
-
-              if (typeof value !== "undefined") {
-                formData.append(name, value);
-              }
-            } catch (error: any) {
-              // Emit error to busboy to bail early if possible
-              busboy.emit("error", error);
-              // It's possible that the handler is doing stuff and fails
-              // *after* busboy has finished. Rethrow the error for surfacing
-              // in the Promise.all(fileWorkQueue) below.
-              throw error;
-            } finally {
-              filestream.resume();
-            }
-          })()
-        );
-      } else {
-        filestream.resume();
-      }
-
-      if (!uploadHandler) {
-        console.warn(
-          `Tried to parse multipart file upload but no uploadHandler was provided.`
-        );
-      }
-    });
-
-    stream.on("error", abort);
-    stream.on("aborted", abort);
-    busboy.on("error", abort);
-    busboy.on("finish", resolve);
-
-    stream.pipe(busboy);
-  });
-
-  await Promise.all(fileWorkQueue);
-
-  return formData;
-}
 
 function isBlob(value: any): value is Blob {
   return (
