+++ packages/remix-node/index.ts
@@ -1,3 +1,5 @@
+export { AbortController } from "abort-controller";
+
 export { formatServerError } from "./errors";
 
 export type {
