+++ packages/remix-vercel/package.json
@@ -13,8 +13,7 @@
   },
   "dependencies": {
     "@remix-run/node": "1.0.6",
-    "@remix-run/server-runtime": "1.0.6",
-    "abort-controller": "^3.0.0"
+    "@remix-run/server-runtime": "1.0.6"
   },
   "peerDependencies": {
     "@vercel/node": "^1.8.3"
